to_float方法改为cast

+++ AboutDL/语音识别/model_language/transformer.py
@@ -288,7 +288,7 @@ class Lm():
             self.logits = tf.layers.dense(self.outputs, self.label_vocab_size)
             self.preds = tf.to_int32(tf.argmax(self.logits, axis=-1))
             self.istarget = tf.cast(tf.not_equal(self.y, 0),tf.float32)
-            self.acc = tf.reduce_sum(tf.to_float(tf.equal(self.preds, self.y))*self.istarget)/ (tf.reduce_sum(self.istarget))
+            self.acc = tf.reduce_sum(tf.cast((tf.equal(self.preds, self.y)),tf.float32)*self.istarget)/ (tf.reduce_sum(self.istarget))
             tf.summary.scalar('acc', self.acc)
                         
             if self.is_training:
